Extend functionality of "image_orientation"
If the image_orientation option is passed when running the PPStructure on an image that is oriented correctly (needs no rotation) the 'cv_rotate_code' will return None and cv2.rotate() will throw an error. By checking if the angle is not 0 before calling cv2.rotate(), this can be avoided.